Fix API request to auto-call ONLY in first render and avoid multiple submit clicks
The free API is restricted to 1000 calls a month, so to make as least calls as possible I check If I already have some info or not.
When the user submits the form, I check in handleSubmit that the call goes out only once previously comparing the stored ip vs. user's input ip.